% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-afs_afr.R
\docType{data}
\name{afs_afr}
\alias{afs_afr}
\title{African/African American target data from gnomAD v2.1 (Karczewski, 2020)}
\format{
A data frame for the Non-Finnish European AFS target data. The first two columns define the MAC Bin boundaries. The third column is the proportion of variants in that bin.
}
\usage{
data(afs_afr)
}
\description{
African/African American target data from gnomAD v2.1 (Karczewski, 2020)
}
\details{
Used to fit the afs function with *fit_afs*
}
\keyword{datasets}
