% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AFS_calc.R
\name{AFS_calc}
\alias{AFS_calc}
\title{RAREsim}
\usage{
AFS_calc(alpha = NULL, beta = NULL, b = NULL, mac, pop = NULL)
}
\arguments{
\item{alpha}{Observed probabilities for each MAC bin}

\item{beta}{Number of individuals in the target data}

\item{b}{percent of rare variants - just the sum of the proportions}

\item{mac}{The MAC bins to use, with lower and upper boundaries defined}

\item{pop}{The population -  specified when using default  parameters}
}
\value{
data frame with the MAC bins provided and proportion of variants in  each bin
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
