% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-afs_eas.R
\docType{data}
\name{afs_eas}
\alias{afs_eas}
\title{East Asian AFS target data from gnomAD (cite)}
\format{
A data frame for the Non-Finnish European AFS target data
}
\source{
RAREsim paper
}
\usage{
data(afs_eas)
}
\description{
East Asian AFS target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
