% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-afs_nfe.R
\docType{data}
\name{afs_nfe}
\alias{afs_nfe}
\title{Non-Finnish European AFS target data from gnomAD (cite)}
\format{
A data frame for the Non-Finnish European AFS target data
}
\source{
RAREsim paper
}
\usage{
data(afs_nfe)
}
\description{
Non-Finnish European AFS target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
