% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-afs_sas.R
\docType{data}
\name{afs_sas}
\alias{afs_sas}
\title{South Asian target data from gnomAD v2.1 (Karczewski, 2020)}
\format{
A data frame for the Non-Finnish European AFS target data. The first two columns define the MAC Bin boundaries. The third column is the proportion of variants in that bin.
}
\usage{
data(afs_sas)
}
\description{
South Asian target data from gnomAD v2.1 (Karczewski, 2020)
}
\details{
Used to fit the afs function with *Fit_afs*
}
\keyword{datasets}
