% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_delete_list.R
\name{create_delete_list}
\alias{create_delete_list}
\title{RAREsim}
\usage{
create_delete_list(variants_to_remove, name = NULL)
}
\arguments{
\item{variants_to_remove}{list of the line  number of variants to remove}

\item{name}{name of the file that will be written}
}
\value{
Text stating what has been accomlished
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
