% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected_variants.R
\name{Expected_variants}
\alias{Expected_variants}
\title{RAREsim}
\usage{
Expected_variants(alpha, beta, b, phi, omega, Ntar, Size)
}
\arguments{
\item{temp}{A data frame with the first column sample size and the second variants per Kb}
}
\value{
data frame with  the bins, proportion, and expected variants
}
\description{
Simulate rare variant genetic data
}
\seealso{
\code{\link{AFS_FitTarget}}
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
