% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_var_per_kb.R
\name{Fit_var_per_kb}
\alias{Fit_var_per_kb}
\title{RAREsim}
\usage{
Fit_var_per_kb(temp)
}
\arguments{
\item{temp}{A data frame with the first column sample size and the second variants per Kb}
}
\value{
Vector of parameters - phi and omega
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
