% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_update.R
\name{m_update}
\alias{m_update}
\title{RAREsim}
\usage{
m_update(con1, con2, repl_lines, line_nos)
}
\arguments{
\item{con1}{original name of the file}

\item{con2}{output file}

\item{repl_lines}{replacement  lines}

\item{line_nos}{line numbers to be  replaces}
}
\value{
Number of variants pruned
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
