% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-MAC_afr.R
\docType{data}
\name{MAC_afr}
\alias{MAC_afr}
\title{Example MAC data for an African/African American population, simulated from HAPGEN2 with all bp included}
\format{
A data frame with one column of MAC's - no column name
}
\usage{
data(MAC_afr)
}
\description{
Example MAC data for an African/African American population, simulated from HAPGEN2 with all bp included
}
\keyword{datasets}
