% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_variant_macs.R
\name{new_variant_macs}
\alias{new_variant_macs}
\title{RAREsim}
\usage{
new_variant_macs(original_variants, new_MAC)
}
\arguments{
\item{new_MAC}{}

\item{original_haps}{}
}
\value{
New dataframe with the edited variants
}
\description{
Edit the necessary variants
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
