% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nvariant_afr.R
\docType{data}
\name{nvariant_afr}
\alias{nvariant_afr}
\title{African target data for the Number of Variants function from gnomADv2.1 (Karczewski, 2020)}
\format{
Number of Variants target data. First column is sample size and second column is number of variants per Kb
}
\usage{
data(nvariant_afr)
}
\description{
African target data for the Number of Variants function from gnomADv2.1 (Karczewski, 2020)
}
\keyword{datasets}
