% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nvariant_eas.R
\docType{data}
\name{nvariant_eas}
\alias{nvariant_eas}
\title{East Asian Variants per Kb target data from gnomADv2.1 (Karczewski, 2020)}
\format{
Number of Variants target data. First column is sample size and second column is number of variants per Kb.
}
\usage{
data(nvariant_eas)
}
\description{
East Asian Variants per Kb target data from gnomADv2.1 (Karczewski, 2020)
}
\keyword{datasets}
