% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nvariant_nfe.R
\docType{data}
\name{nvariant_nfe}
\alias{nvariant_nfe}
\title{Non-Finnish European target data from gnomAD (Karczewski, 2020)}
\format{
Number of Variants target data. First column is sample size and second column is number of variants per Kb.
}
\usage{
data(nvariant_nfe)
}
\description{
Non-Finnish European target data from gnomAD (Karczewski, 2020)
}
\keyword{datasets}
