% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-nvariant_sas.R
\docType{data}
\name{nvariant_sas}
\alias{nvariant_sas}
\title{South Asian Variants per  Kb target data from gnomAD (Karczewski, 2020)}
\format{
Number of Variants target data. First column is sample size and second column is number of variants per Kb
}
\usage{
data(nvariant_sas)
}
\description{
South Asian Variants per  Kb target data from gnomAD (Karczewski, 2020)
}
\keyword{datasets}
