% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning_function.R
\name{Prune}
\alias{Prune}
\title{RAREsim}
\usage{
Prune(hap_file_name, MAC, expected)
}
\arguments{
\item{hap_file_name}{Name of the haplotype file to prune (gzipped: end in .gz)}

\item{MAC}{vector of allele count per variant that corresponds with the haplotype file}

\item{expected}{dataframe of expected number of variants per bin - as well as bin boundaries}
}
\value{
Number of variants pruned
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
