% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_variants.R
\name{prune_variants}
\alias{prune_variants}
\title{Theoretically prune the variants}
\usage{
prune_variants(MAC, expected)
}
\arguments{
\item{MAC}{vector of allele count per variant that corresponds with the haplotype file. 
Line k of the MAC file needs to correspond with the variant on line k in the haplotype/legend files}

\item{expected}{dataframe of expected number of variants per bin - as well as bin boundaries}
}
\value{
A list with two data frames: ToRemove (variants that will have all alternate alleles pruned) and ToChange (variants that will have a subset of alternate alleles removed), Both have 3 columns: line number, current MAC, new MAC
}
\description{
This function theoretically prunes the variants to match what is expected
The current allele counts from the haplotype file also needs to be provided
The output will be used to prune the haplotype files, as seen in the example on github.
}
