% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning_function.R
\name{Pruning_function}
\alias{Pruning_function}
\title{RAREsim}
\usage{
Pruning_function(hap_file_name, MAC, expected, sed_or_gsed)
}
\arguments{
\item{hap_file_name}{Name of the haplotype file to prune (gzipped: end in .gz)}

\item{MAC}{vector of allele count per variant that corresponds with the haplotype file}

\item{expected}{dataframe of expected number of variants per bin - as well as bin boundaries}

\item{sed_or_gsed}{Specify if gsed is loaded onto computer instead of sed}
}
\value{
Number of variants pruned
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
