% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning_function2.R
\name{Pruning_function2}
\alias{Pruning_function2}
\title{RAREsim}
\usage{
Pruning_function2(hap_file_name, MAC, expected, computer_type)
}
\arguments{
\item{hap_file_name}{Name of the haplotype file to prune (gzipped: end in .gz)}

\item{MAC}{vector of allele count per variant that corresponds with the haplotype file}

\item{expected}{dataframe of expected number of variants per bin - as well as bin boundaries}

\item{computer_type}{Specify is a mac is being used}
}
\value{
Number of variants pruned
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
