% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pruning_info.R
\name{Pruning_info}
\alias{Pruning_info}
\title{Pruning Info}
\usage{
Pruning_info(MAC, expected)
}
\arguments{
\item{MAC}{vector of allele count per variant that corresponds with the haplotype file. 
Line k of the MAC file needs to correspond with the variant on line k in the haplotype/legend files}

\item{expected}{dataframe of expected number of variants per bin - as well as bin boundaries}
}
\value{
ToPrune -  a  data frame with 3 columns: line number, current MAC, new MAC
}
\description{
This function theoretically prunes the variants to match what is expected
The current allele counts from the haplotype file also needs to be provided
The output will be used to prune the haplotype files, as seen in the example on github.
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
