% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tot_num_var_afr.R
\docType{data}
\name{tot_num_var_afr}
\alias{tot_num_var_afr}
\title{African target data from gnomAD (cite)}
\format{
A data frame for the African Total Number of Variants target data
}
\source{
RAREsim paper
}
\usage{
data(tot_num_var_afr)
}
\description{
African target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
