% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tot_num_var_eas.R
\docType{data}
\name{tot_num_var_eas}
\alias{tot_num_var_eas}
\title{East Asian target data from gnomAD (cite)}
\format{
A data frame for the East Asian Total Number of Variants target data
}
\source{
RAREsim paper
}
\usage{
data(tot_num_var_eas)
}
\description{
East Asian target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
