% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tot_num_var_nfe.R
\docType{data}
\name{tot_num_var_nfe}
\alias{tot_num_var_nfe}
\title{Non-Finnish European target data from gnomAD (cite)}
\format{
A data frame for the Non-Finnish European Total Number of Variants target data
}
\source{
RAREsim paper
}
\usage{
data(tot_num_var_nfe)
}
\description{
Non-Finnish European target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
