% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-tot_num_var_sas.R
\docType{data}
\name{tot_num_var_sas}
\alias{tot_num_var_sas}
\title{South Asian target data from gnomAD (cite)}
\format{
A data frame for the South Asian Total Number of Variants target data
}
\source{
RAREsim paper
}
\usage{
data(tot_num_var_sas)
}
\description{
South Asian target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
