% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_number_variants.R
\name{Total_number_variants}
\alias{Total_number_variants}
\title{RAREsim}
\usage{
Total_number_variants(phi, omega, n)
}
\arguments{
\item{phi}{parameter phi}

\item{omega}{parameter omega}

\item{n}{sample size in number of individuals}
}
\value{
the number of variants per kb
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
