% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Total_Num_Var_FitTarget.R
\name{Total_variants}
\alias{Total_variants}
\title{RAREsim}
\usage{
Total_variants(temp)
}
\arguments{
\item{temp}{A data frame with the first column sample size and the second variants per Kb}
}
\value{
Vector of parameters - phi and omega
}
\description{
Simulate rare variant genetic data
}
\seealso{
\code{\link{AFS_FitTarget}}
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
