% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-var_per_kb_afr.R
\docType{data}
\name{var_per_kb_afr}
\alias{var_per_kb_afr}
\title{African target data for the Variants per Kb function from gnomADv2.1}
\format{
A data frame for the African Variants per Kb target data
}
\source{
RAREsim paper
}
\usage{
data(var_per_kb_afr)
}
\description{
African target data for the Variants per Kb function from gnomADv2.1
}
\references{
Null, Megan M
}
\keyword{datasets}
