% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-var_per_kb_eas.R
\docType{data}
\name{var_per_kb_eas}
\alias{var_per_kb_eas}
\title{East Asian Variants per Kb target data from gnomADv2.1 (cite)}
\format{
A data frame for the East Asian Variants per Kb target data
}
\source{
RAREsim paper
}
\usage{
data(var_per_kb_eas)
}
\description{
East Asian Variants per Kb target data from gnomADv2.1 (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
