% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-var_per_kb_nfe.R
\docType{data}
\name{var_per_kb_nfe}
\alias{var_per_kb_nfe}
\title{Non-Finnish European target data from gnomAD (cite)}
\format{
A data frame for the Non-Finnish European Variants per Kb target data
}
\source{
RAREsim paper
}
\usage{
data(var_per_kb_nfe)
}
\description{
Non-Finnish European target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
