% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-var_per_kb_sas.R
\docType{data}
\name{var_per_kb_sas}
\alias{var_per_kb_sas}
\title{South Asian Variants per  Kb target data from gnomAD (cite)}
\format{
A data frame for the South Asian Variants per Kb target data
}
\source{
RAREsim paper
}
\usage{
data(var_per_kb_sas)
}
\description{
South Asian Variants per  Kb target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
