% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/afs.R
\name{afs}
\alias{afs}
\title{afs function - Calculates proportion of variants per rare MAC bins}
\usage{
afs(alpha = NULL, beta = NULL, b = NULL, mac_bins, pop = NULL)
}
\arguments{
\item{alpha}{AFS function parameter alpha, does not need to be specified if default parameters are used}

\item{beta}{AFS function parameter beta, does not need to be specified if default parameters are used}

\item{b}{AFS function parameter b, does not need to be specified if default parameters are used}

\item{mac_bins}{The MAC bins to use, with lower and upper boundaries defined}

\item{pop}{The population: AFR, EAS, NFE or SAS - specified when using default parameters}
}
\value{
data frame with the MAC bins provided and proportion of variants in each bin
}
\description{
The afs function calculates the proportion of variants in minor allele count (MAC) bins given parameters, as described in RAREsim.
The default parameters will be used if an ancestrial population is specified: (AFR, EAS, NFE, or SAS) is specified.
Otherwise, the parameters alpha, beta, and b need to be provided.
Alpha, Beta, and b can be estimated from target data using the *Fit_afs* function.
}
\details{
The default parameters will be used if an ancestrial population is specified: (AFR, EAS, NFE, or SAS) is specified.
Otherwise, the parameters alpha, beta, and b need to be provided.
Alpha, Beta, and b can be estimated from target data using the *Fit_afs* function.
}
