% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-afs_afr.R
\docType{data}
\name{afs_afr}
\alias{afs_afr}
\title{African/African American target data from gnomAD (cite)}
\format{
A data frame for the Non-Finnish European AFS target data
}
\source{
RAREsim paper
}
\usage{
data(afs_afr)
}
\description{
African/African American target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
