% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-afs_sas.R
\docType{data}
\name{afs_sas}
\alias{afs_sas}
\title{South Asian target data from gnomAD (cite)}
\format{
A data frame for the Non-Finnish European AFS target data
}
\source{
RAREsim paper
}
\usage{
data(afs_sas)
}
\description{
South Asian target data from gnomAD (cite)
}
\references{
Null, Megan M
}
\keyword{datasets}
