% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_delete_list.R
\name{create_delete_list}
\alias{create_delete_list}
\title{Create file for effiecient pruning of variants}
\usage{
create_delete_list(variants_to_remove, name = NULL)
}
\arguments{
\item{variants_to_remove}{list of the line  number of variants to remove}

\item{name}{name of the file that will be written. Default to 'List2delete.sed'}
}
\value{
R output is text stating what has been accomlished. A file has been written.
}
\description{
Create file for effiecient pruning of variants
}
