% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected_variants.R
\name{Expected_variants}
\alias{Expected_variants}
\title{RAREsim}
\usage{
Expected_variants(
  alpha = NULL,
  beta = NULL,
  b = NULL,
  phi = NULL,
  omega = NULL,
  N,
  Size,
  pop = NULL,
  mac
)
}
\arguments{
\item{alpha}{AFS funciton parameter alpha}

\item{beta}{AFS funciton parameter beta}

\item{b}{AFS funciton parameter b}

\item{phi}{Variants per Kb funciton parameter phi}

\item{omega}{Variants per Kb funciton parameter omega}

\item{N}{sample size to simulate}

\item{Size}{Simulation region sample size in Kb}

\item{pop}{Ancestry population - needs to be specified ('AFR', 'NFE', 'EAS', or 'SAS') if using default}

\item{mac}{The MAC bins to use, with lower and upper boundaries defined. Only define for rare variants}
}
\value{
data frame with the MAC bins and expected variants
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
