% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected_variants.R
\name{Expected_variants}
\alias{Expected_variants}
\title{RAREsim}
\usage{
Expected_variants(alpha, beta, b, phi, omega, Ntar, Size)
}
\arguments{
\item{alpha}{parameter from AFS funciton}

\item{beta}{another parameter from AFS funciton}

\item{b}{final parameter from AFS funciton}

\item{phi}{parameter from total variants funciton}

\item{omega}{other parameter from total variants funciton}

\item{Ntar}{sample size of the target data}

\item{Size}{Simulation region sample size in Kb}
}
\value{
data frame with  the bins, proportion, and expected variants
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
