% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Expected_variants.R
\name{Expected_variants}
\alias{Expected_variants}
\title{Expected Variants}
\usage{
Expected_variants(Total_num_var, mac_bin_prop)
}
\arguments{
\item{Total_num_var}{estimated  number of variants  in the  region}

\item{mac_bin_prop}{The MAC bins to use, with lower and upper boundaries defined. Only define for rare variants}
}
\value{
data frame with the MAC bins and expected variants
}
\description{
This function combines the Number of Variants and AFS functions to produces the expected number of variants per Kb in each MAC bin
The output from the afs and nvariants functions can be used here
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
