% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_AFS.R
\name{Fit_AFS}
\alias{Fit_AFS}
\title{RAREsim}
\usage{
Fit_AFS(prop, N, per_rv)
}
\arguments{
\item{prop}{Observed probabilities for each MAC bin}

\item{N}{Number of individuals in the target data}

\item{per_rv}{percent of rare variants - just the sum of the proportions}
}
\value{
Vector of parameters - alpha, beta, and b
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
