% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fit_afs.R
\name{Fit_afs}
\alias{Fit_afs}
\title{RAREsim}
\usage{
Fit_afs(prop_df, N, p_rv = NULL)
}
\arguments{
\item{prop_df}{data frame with 3 columns, Lower, Upper (of MAC bin)  and proportion of variants in that MAC bin}

\item{N}{Number of individuals in the target data}

\item{p_rv}{percent of rare variants - just the sum of the proportions}
}
\value{
Vector of parameters - alpha, beta, and b as well as  fitted values
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
