% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nvariant.R
\name{fit_nvariant}
\alias{fit_nvariant}
\title{Given target data, fit the Number of Variants function}
\usage{
fit_nvariant(Observed_variants_per_kb)
}
\arguments{
\item{Observed_variants_per_kb}{A data frame with the first column sample size and the second variants per Kb, both numeric}
}
\value{
Vector of parameters - phi and omega
}
\description{
This function takes Number of Variants target data and estimates parameters for the Number of Variants function.
A dataframe specifying the number of variants per Kb at various sample sizes is required to fit the data
}
\examples{
data("nvariant_afr")
fit_nvariant(nvariant_afr)

}
