% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-MAC_afr.R
\docType{data}
\name{MAC_afr}
\alias{MAC_afr}
\title{African example MAC data, simulated from HAPGEN2 with all bp included}
\format{
A data frame with one column of MAC's
}
\source{
RAREsim paper
}
\usage{
data(MAC_afr)
}
\description{
African example MAC data, simulated from HAPGEN2 with all bp included
}
\references{
Null, Megan M
}
\keyword{datasets}
