% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_variant_macs.R
\name{new_variant_macs}
\alias{new_variant_macs}
\title{Edits the haplotype file for variants that require a subset of alternate alleles returned to reference}
\usage{
new_variant_macs(original_variants, new_MAC)
}
\arguments{
\item{original_variants}{A data frame with the haplotypes of the variants the require a subset of the alternate alleles to be returned to reference}

\item{new_MAC}{The new MAC for each variant. The variants counts need to match the line number of original_haps.}
}
\value{
New dataframe with the edited variants to insert into the complete haplotype files
}
\description{
The *new_variant_macs* function edits the variants that require a subset of alternate alleles returned to reference.
The variants that will be edited should first extracted from the haplotype file and the new MAC for each variant determined.
See the example scripts for how to extract the variants, and the *pruning_info* function to determine which
variants should have a subset of alternate alleles returned to reference.
}
