% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nvariant.R
\name{nvariant}
\alias{nvariant}
\title{RAREsim}
\usage{
nvariant(phi = NULL, omega = NULL, n, pop = NULL)
}
\arguments{
\item{phi}{parameter phi}

\item{omega}{parameter omega}

\item{n}{sample size in number of individuals}

\item{pop}{population - only needs to be specified if  using default  parameters}
}
\value{
the number of variants per kb
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
