% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hap_files.R
\name{update_hap_files}
\alias{update_hap_files}
\title{RAREsim}
\usage{
update_hap_files(con1, con2, repl_lines, line_nos)
}
\arguments{
\item{con1}{original name of the file}

\item{con2}{output file}

\item{repl_lines}{replacement lines - obtained from readLines of the file to replace}

\item{line_nos}{line numbers to be replaced}
}
\value{
R output is text stating what has been accomplished. The new files have been written.
}
\description{
Simulate rare variant genetic data
}
\author{
Megan M Null, \email{megan.null@ucdenver.edu}
}
\keyword{RAREsim}
